Implementing local permission system

+++ project3.py
@@ -1,8 +1,7 @@
-from flask import Flask, render_template, request, redirect, jsonify, url_for, \
-    flash
+from flask import Flask, render_template, request, redirect, url_for, flash
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Music, Users, base
+from database_setup import User, base, Playlist
 from flask import session as login_session
 import random
 import string
@@ -27,19 +26,21 @@ db_session = sessionmaker(bind=engine)
 session = db_session()
 
 
-# Create anti-forgery state token
 @app.route('/login')
 def login():
+    """ Encodes login URL for anti-forgery state token. """
+
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
                     for x in xrange(32))
     login_session['state'] = state
-    # return "The current session state is {}" .format login_session['state']
     return render_template('login.html', STATE=state)
 
 
 @app.route('/gconnect', methods=['POST'])
 def g_connect():
-    # Validate state token
+    """ Connects and authorizes user against Google's Google+ API. """
+
+    # Validating state token
     if request.args.get('state') != login_session['state']:
         response = make_response(json.dumps('Invalid state parameter.'), 401)
         response.headers['Content-Type'] = 'application/json'
@@ -52,6 +53,7 @@ def g_connect():
         oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
         oauth_flow.redirect_uri = 'postmessage'
         credentials = oauth_flow.step2_exchange(code)
+
     except FlowExchangeError:
         response = make_response(
             json.dumps('Failed to upgrade the authorization code.'), 401)
@@ -64,6 +66,7 @@ def g_connect():
            '}'.format(access_token))
     h = httplib2.Http()
     result = json.loads(h.request(url, 'GET')[1])
+
     # If there was an error in the access token info, abort.
     if result.get('error') is not None:
         response = make_response(json.dumps(result.get('error')), 500)
@@ -102,12 +105,16 @@ def g_connect():
     user_info_url = "https://www.googleapis.com/oauth2/v1/userinfo"
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(user_info_url, params=params)
-
     data = answer.json()
 
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
+    #
+    user_id = get_user_id(login_session['email'])
+    if not user_id:
+        create_user(login_session)
+    login_session['user_id'] = user_id
 
     output = ''
     output += '<h1>Welcome, '
@@ -115,17 +122,20 @@ def g_connect():
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;\
+        -webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as {}".format(login_session['username']))
     print "done!"
     return output
 
-    # DISCONNECT - Revoke a current user's token and reset their login_session
-
 
 @app.route('/gdisconnect')
 def g_disconnect():
-    # Only disconnect a connected user.
+    """
+    Revoke a current user's token and reset their login_session
+    Only disconnect a connected user.
+    """
+
     credentials = login_session.get('credentials')
     if credentials is None:
         response = make_response(
@@ -138,17 +148,17 @@ def g_disconnect():
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
 
+    # Reset the user's session.
     if result['status'] == '200':
-        # Reset the user's sesson.
         del login_session['credentials']
         del login_session['gplus_id']
         del login_session['username']
         del login_session['email']
         del login_session['picture']
-
         response = make_response(json.dumps('Successfully disconnected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
+
     else:
         # For whatever reason, the given token was invalid.
         response = make_response(
@@ -157,36 +167,58 @@ def g_disconnect():
         return response
 
 
-@app.route('/', methods=['GET', 'POST'])
-@app.route('/users/', methods=['GET', 'POST'])
+@app.route('/')
+@app.route('/users/')
 def home():
-    users = session.query(Users).order_by(asc(Users.name))
+    """ Renders the home template and passes on a list of all users and their
+     names
+    """
+
+    users = session.query(User).order_by(asc(User.name))
     return render_template('home.html', users=users)
 
 
 @app.route('/user/<int:user_id>/')
 def show_user(user_id):
-    users = session.query(Users).filter_by(id=user_id)
+    users = session.query(User).filter_by(id=user_id)
     return render_template('show-user.html', user_id=user_id, users=users)
 
 
-@app.route('/create/', methods=['GET', 'POST'])
-def create_user():
-    if request.method == 'POST':
-        user_to_create = Users(name=request.form['name'],
-                               email=request.form['email'])
-        session.add(user_to_create)
-        session.commit()
-        flash(
-            'The user {} was successfully created'.format(user_to_create.name))
-        return redirect(url_for('home'))
-    else:
-        return render_template('create-user.html')
+def get_user_id(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return
+
+
+def get_user_info(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def create_user(login_session):
+    new_user = User(name=login_session['username'],
+                    email=login_session['email'],
+                    picture=login_session['picture'])
+    session.add(new_user)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def create_playlist():
+    new_playlist = Playlist(name=request.form['name'],
+                            description=request.form['description'],
+                            user_id=User.id)
+    session.add(new_playlist)
+    session.commit()
+    return redirect(url_for('show_user'))
 
 
 @app.route('/user/<int:user_id>/edit/', methods=['GET', 'POST'])
 def edit_user(user_id):
-    user_to_edit = session.query(Users).filter_by(id=user_id).one()
+    user_to_edit = session.query(User).filter_by(id=user_id).one()
     if request.method == 'POST':
         if request.form['name']:
             user_to_edit.name = request.form['name']
@@ -197,13 +229,17 @@ def edit_user(user_id):
         flash('User {} updated successfully'.format(user_to_edit.name))
         return redirect(url_for('show_user', user_id=user_id))
     else:
-        return render_template('edit-user.html', user_id=user_id,
+        return render_template('edit-user.html',
+                               user_id=user_id,
                                user_to_edit=user_to_edit)
 
 
+# FIXME: Clicking delete button doesn't trigger a POST request to delete user
 @app.route('/user/<int:user_id>/delete/', methods=['GET', 'POST'])
 def delete_user(user_id):
-    user_to_delete = session.query(Users).filter_by(id=user_id).one()
+    user_to_delete = session.query(User).filter_by(id=user_id).one()
+    if 'username' not in login_session:
+        return redirect('login')
     if request.method == 'POST':
         session.delete(user_to_delete)
         session.commit()
